backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-docx==1.1.0
pdfplumber==0.10.3
openai==1.12.0
openai-whisper==20231117
rapidfuzz==3.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
